Fix XML first line generation
