+++ drafthorse/__init__.py
@@ -1,1 +1,1 @@
-version = "2.2.2"
+version = "2.2.4"
